refactor: 아이디 중복 로직 수정
